Rework concept of reserved words.
For each entity that has reserved words or special mapping of attribute names (e.g. name becomes repositoryName), we create a plist named EntityNameAttributeMap. Within this file, the key represents the local attribute name, and the value is the corresponding remote attribute (repositoryName:name in the above example). Support for this is also included when setting the predicateKey variable.